Moved test_utils to it's own directory out of core. Updated references (#2154)
