+++ osquery/config/parsers/tests/events_tests.cpp
@@ -14,7 +14,7 @@
 #include <osquery/config.h>
 #include <osquery/registry.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 
 namespace osquery {
