+++ osquery/config/parsers/tests/options_tests.cpp
@@ -13,7 +13,7 @@
 #include <osquery/config.h>
 #include <osquery/registry.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 
 namespace osquery {
