+++ osquery/config/tests/config_tests.cpp
@@ -22,7 +22,7 @@
 #include <osquery/registry.h>
 #include <osquery/sql.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 
 namespace pt = boost::property_tree;
