+++ osquery/core/tests/conversions_tests.cpp
@@ -18,7 +18,7 @@
 
 #include "osquery/core/conversions.h"
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 
 #define EXPECT_WITHIN_INCLUSIVE(lower, upper, val)                     \
   do {                                                                 \
