+++ osquery/core/tests/hash_tests.cpp
@@ -12,7 +12,7 @@
 
 #include <osquery/hash.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 
 namespace osquery {
