+++ osquery/core/tests/permissions_tests.cpp
@@ -21,7 +21,7 @@
 #include <osquery/core.h>
 #include <osquery/logger.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 
 namespace fs = boost::filesystem;
