+++ osquery/database/benchmarks/database_benchmarks.cpp
@@ -13,7 +13,7 @@
 #include <osquery/database.h>
 #include <osquery/filesystem.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 #include "osquery/database/query.h"
 
 namespace osquery {
