+++ osquery/database/tests/results_tests.cpp
@@ -17,7 +17,7 @@
 #include <osquery/database.h>
 #include <osquery/logger.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 
 namespace pt = boost::property_tree;
