+++ osquery/dispatcher/tests/scheduler_tests.cpp
@@ -12,7 +12,7 @@
 
 #include <osquery/logger.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 #include "osquery/sql/sqlite_util.h"
 #include "osquery/dispatcher/scheduler.h"
