+++ osquery/distributed/tests/distributed_tests.cpp
@@ -20,7 +20,7 @@
 #include <osquery/enroll.h>
 #include <osquery/sql.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 
 #include "osquery/sql/sqlite_util.h"
