+++ osquery/filesystem/darwin/benchmarks/plist_benchmarks.cpp
@@ -15,7 +15,7 @@
 #include <osquery/filesystem.h>
 
 #include "osquery/core/conversions.h"
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 
 namespace fs = boost::filesystem;
 namespace pt = boost::property_tree;
