+++ osquery/filesystem/darwin/tests/plist_tests.cpp
@@ -16,7 +16,7 @@
 #include <osquery/logger.h>
 
 #include "osquery/core/conversions.h"
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 
 namespace fs = boost::filesystem;
 namespace pt = boost::property_tree;
