+++ osquery/logger/plugins/tests/aws_firehose_tests.cpp
@@ -20,7 +20,7 @@
 
 #include <osquery/logger.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 #include "osquery/logger/plugins/aws_firehose.h"
 #include "osquery/logger/plugins/aws_util.h"
