+++ osquery/logger/plugins/tests/aws_kinesis_tests.cpp
@@ -18,7 +18,7 @@
 
 #include <osquery/logger.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 #include "osquery/logger/plugins/aws_kinesis.h"
 #include "osquery/logger/plugins/aws_util.h"
