+++ osquery/logger/plugins/tests/aws_util_tests.cpp
@@ -14,7 +14,7 @@
 
 #include <osquery/logger.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 #include "osquery/logger/plugins/aws_util.h"
 
 namespace osquery {
