+++ osquery/logger/plugins/tests/buffered_tests.cpp
@@ -20,7 +20,7 @@
 #include <osquery/dispatcher.h>
 #include <osquery/logger.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 #include "osquery/logger/plugins/buffered.h"
 
 using namespace testing;
