+++ osquery/logger/plugins/tests/filesystem_logger_tests.cpp
@@ -14,7 +14,7 @@
 
 #include <osquery/logger.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 
 namespace fs = boost::filesystem;
