+++ osquery/logger/plugins/tests/tls_logger_tests.cpp
@@ -13,7 +13,7 @@
 #include <osquery/logger.h>
 #include <osquery/database.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 
 #include "osquery/logger/plugins/tls.h"
