+++ osquery/main/tests.cpp
@@ -21,7 +21,7 @@
 #include <osquery/logger.h>
 
 #include "osquery/core/process.h"
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 #include "osquery/core/testing.h"
 
 namespace osquery {
