+++ osquery/remote/enroll/tests/enroll_tests.cpp
@@ -16,7 +16,7 @@
 #include <osquery/filesystem.h>
 #include <osquery/flags.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 
 namespace osquery {
