+++ osquery/remote/transports/tests/tls_transports_tests.cpp
@@ -18,7 +18,7 @@
 #include "osquery/remote/serializers/json.h"
 #include "osquery/remote/transports/tls.h"
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 
 namespace pt = boost::property_tree;
