+++ osquery/sql/tests/sqlite_util_tests.cpp
@@ -15,7 +15,7 @@
 #include <osquery/core.h>
 #include <osquery/sql.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 #include "osquery/sql/sqlite_util.h"
 
 namespace osquery {
