+++ osquery/tables/networking/darwin/tests/wifi_tests.mm
@@ -12,7 +12,7 @@
 #include <Foundation/Foundation.h>
 #include <gtest/gtest.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 
 namespace osquery {
 namespace tables {
