+++ osquery/tables/networking/linux/tests/iptables_tests.cpp
@@ -15,7 +15,7 @@
 #include <libiptc/libiptc.h>
 #include <arpa/inet.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 
 namespace osquery {
 namespace tables {
