+++ osquery/tables/networking/tests/networking_tables_tests.cpp
@@ -12,7 +12,7 @@
 
 #include <osquery/logger.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 
 namespace osquery {
 namespace tables {
