+++ osquery/tables/system/darwin/tests/apps_tests.cpp
@@ -14,7 +14,7 @@
 #include <osquery/logger.h>
 #include <osquery/tables.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 
 namespace fs = boost::filesystem;
 namespace pt = boost::property_tree;
