+++ osquery/tables/system/darwin/tests/asl_tests.cpp
@@ -20,7 +20,7 @@
 #include <osquery/sql.h>
 
 #include "osquery/tables/system/darwin/asl_utils.h"
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 
 namespace pt = boost::property_tree;
