+++ osquery/tables/system/darwin/tests/certificates_tests.cpp
@@ -13,7 +13,7 @@
 #include <osquery/logger.h>
 
 #include "osquery/tables/system/darwin/keychain.h"
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 
 namespace osquery {
 namespace tables {
