+++ osquery/tables/system/darwin/tests/extended_attributes_tests.cpp
@@ -16,7 +16,7 @@
 
 #include <osquery/filesystem.h>
 #include <osquery/logger.h>
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 
 namespace osquery {
 namespace tables {
