+++ osquery/tables/system/darwin/tests/launchd_tests.cpp
@@ -13,7 +13,7 @@
 #include <osquery/filesystem.h>
 #include <osquery/logger.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 
 namespace fs = boost::filesystem;
 namespace pt = boost::property_tree;
