+++ osquery/tables/system/darwin/tests/signature_tests.mm
@@ -15,7 +15,7 @@
 #include <boost/make_unique.hpp>
 #include <mach-o/dyld.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 
 namespace fs = boost::filesystem;
