+++ osquery/tables/system/darwin/tests/smc_tests.cpp
@@ -13,7 +13,7 @@
 #include <osquery/logger.h>
 #include <osquery/tables.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 
 namespace osquery {
 namespace tables {
