+++ osquery/tests/test_util.cpp
@@ -26,7 +26,7 @@
 #include <osquery/logger.h>
 #include <osquery/sql.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 
 namespace fs = boost::filesystem;
