+++ tools/tests/plist_benchmark.cpp
@@ -15,7 +15,7 @@
 #include <osquery/flags.h>
 #include <osquery/logger.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/tests/test_util.h"
 
 namespace pt = boost::property_tree;
